content/canvas/test/test_2d.path.isPointInPath.unclosed.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.path.isPointInPath.unclosed</title>
<!-- Testing: isPointInPath() works on unclosed subpaths -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.moveTo(0, 0);
ctx.lineTo(20, 0);
ctx.lineTo(20, 20);
ctx.lineTo(0, 20);
ok(ctx.isPointInPath(10, 10) === true, "ctx.isPointInPath(10, 10) === true");
ok(ctx.isPointInPath(30, 10) === false, "ctx.isPointInPath(30, 10) === false");

SimpleTest.finish();

});
</script>
